+++ src/main/java/fr/insee/genesis/domain/ports/api/RawResponseApiPort.java
@@ -24,10 +24,11 @@ public interface RawResponseApiPort {
     List<String> getUnprocessedCollectionInstrumentIds();
     void updateProcessDates(List<SurveyUnitModel> surveyUnitModels);
     Page<RawResponseModel> findRawResponseDataByCampaignIdAndDate(String campaignId, Instant startDate, Instant endDate, Pageable pageable);
+    long countDistinctInterrogationIdsByCollectionInstrumentId(String collectionInstrumentId);
 
     long countByCollectionInstrumentId(String collectionInstrumentId);
 
     Set<String> getDistinctCollectionInstrumentIds();
     Page<RawResponseModel> findRawResponseDataByCollectionInstrumentId(String collectionInstrumentId, Pageable pageable);
-
+    boolean existsByInterrogationId(String interrogationId);
 }