+++ CHANGELOG.md
@@ -1,4 +1,29 @@
 # Changelog
+## 2.4.0 [2026-03-04]
+### Added
+- Restful endpoints to get latest responses without the logic in it (not to be confused with get latest states)
+- Simplified survey units now have datastate in their variables
+- Volumetrics with distinct interrogationId
+- 
+### Changed
+- Deprecated non-restful old endpoint
+### Fixed
+- Returned null collectionInstrumentId in get unprocessed
+
+## 2.3.1 [2026-02-24]
+### Fixed
+- 500 when save edited without campaignId
+
+## 2.3.0 [2026-02-19]
+### Added
+- Get response by-collection-instrument-and-between-datetime endpoint
+### Fixed
+- Get modes by question collectionInstrumentId parameter
+
+## 2.2.3 [2026-02-12]
+### Fixed
+- Fixed INTEGER parsing to use Long instead of int to support large values
+
 ## 2.2.2 [2026-02-05]
 ### Fixed
 - Volumetry don't count responses with collectionInstrumentId